Add a "bounded runtime" mode to guest execution.
This modifies the instruction-counting mechanism to track an
instruction-count bound as well; and when a bound is set, the Wasm guest
will make a hostcall to yield back to the host context.

This is all placed under the `run_async()` function, so bounded-execution
yields manifest as futures that are immediately ready to continue execution.
This should give an executor main loop a chance to do other work at regular
intervals.